consolidate draft mode tests
